feat(scripts,docs,skills): add Microsoft Learn MCP + scripts refactor
- Add Microsoft Learn MCP server (.vscode/mcp.json, agent registry, docs)
- Add 3 new skills: microsoft-docs, microsoft-code-reference, microsoft-skill-creator
- Update all docs references from 4 → 5 MCP servers (six files)
- Add Microsoft Learn row to four-pillars.md MCP table

scripts refactor (net -224 lines, 40 files):
- Delete validate-agent-body-size.mjs and validate-skill-size.mjs (merged into
  validate-agent-frontmatter and validate-skills-format respectively)
- Extract ARTIFACT_HEADINGS to scripts/_lib/artifact-headings.mjs (single source
  of truth); update validate-artifact-templates and fix-artifact-h2 to import from it
- Adopt Reporter class in validate-instruction-frontmatter, validate-version-sync,
  validate-docs-sync, validate-mcp-config, validate-agent-frontmatter,
  validate-skills-format
- Fix workspace-index.mjs to import path constants from paths.mjs (no duplicates)
- Fix validate-skill-digests to use getSkills() from workspace-index
- Fix validate-h2-sync to read ARTIFACT_HEADINGS from new _lib/ location
- Add validate:docs-sync and validate:vscode to validate:_node parallel runner
- Remove 3 dead npm script aliases from package.json
- Parallelize diff-based-push-check.sh with background jobs + temp dir result files
- Fix dead links in docs/CHANGELOG.md and docs/CONTRIBUTING.md (docs/ copies)

ci: remove failing auto-merge step from azure-deprecation-tracker.yml
  (GraphQL: Pull request Auto merge is not allowed for this repository)
