+++ .github/agents/05b-bicep-planner.agent.md
@@ -51,6 +51,7 @@ tools:
     web/fetch,
     web/githubRepo,
     "azure-mcp/*",
+    "microsoft-learn/*",
     "bicep/*",
     todo,
     vscode.mermaid-chat-features/renderMermaidDiagram,
