+++ .github/agents/05t-terraform-planner.agent.md
@@ -51,6 +51,7 @@ tools:
     web/fetch,
     web/githubRepo,
     "azure-mcp/*",
+    "microsoft-learn/*",
     "terraform/*",
     todo,
     vscode.mermaid-chat-features/renderMermaidDiagram,
