+++ .github/skill-affinity.json
@@ -10,7 +10,7 @@
         "azure-artifacts",
         "workflow-engine"
       ],
-      "secondary": ["context-shredding", "github-operations"],
+      "secondary": ["context-shredding", "github-operations", "microsoft-docs"],
       "never": [
         "azure-bicep-patterns",
         "terraform-patterns",
@@ -20,7 +20,7 @@
     },
     "02-Requirements": {
       "primary": ["azure-defaults", "azure-artifacts", "session-resume"],
-      "secondary": [],
+      "secondary": ["microsoft-docs"],
       "never": [
         "azure-bicep-patterns",
         "terraform-patterns",
@@ -31,12 +31,12 @@
     },
     "03-Architect": {
       "primary": ["azure-defaults", "azure-artifacts", "session-resume"],
-      "secondary": ["context-shredding", "azure-diagrams"],
+      "secondary": ["context-shredding", "azure-diagrams", "microsoft-docs"],
       "never": ["azure-bicep-patterns", "terraform-patterns", "iac-common"]
     },
     "04-Design": {
       "primary": ["azure-diagrams", "azure-adr"],
-      "secondary": ["azure-artifacts", "session-resume"],
+      "secondary": ["azure-artifacts", "session-resume", "microsoft-docs"],
       "never": [
         "azure-bicep-patterns",
         "terraform-patterns",
@@ -51,7 +51,11 @@
         "azure-bicep-patterns",
         "session-resume"
       ],
-      "secondary": ["context-shredding"],
+      "secondary": [
+        "context-shredding",
+        "microsoft-docs",
+        "microsoft-code-reference"
+      ],
       "never": ["terraform-patterns", "azure-troubleshooting", "iac-common"]
     },
     "05t-Terraform Planner": {
@@ -61,7 +65,11 @@
         "terraform-patterns",
         "session-resume"
       ],
-      "secondary": ["context-shredding"],
+      "secondary": [
+        "context-shredding",
+        "microsoft-docs",
+        "microsoft-code-reference"
+      ],
       "never": ["azure-bicep-patterns", "azure-troubleshooting", "iac-common"]
     },
     "06b-Bicep CodeGen": {
@@ -71,7 +79,11 @@
         "azure-bicep-patterns",
         "session-resume"
       ],
-      "secondary": ["context-shredding"],
+      "secondary": [
+        "context-shredding",
+        "microsoft-docs",
+        "microsoft-code-reference"
+      ],
       "never": ["terraform-patterns", "azure-troubleshooting", "iac-common"]
     },
     "06t-Terraform CodeGen": {
@@ -81,7 +93,11 @@
         "terraform-patterns",
         "session-resume"
       ],
-      "secondary": ["context-shredding"],
+      "secondary": [
+        "context-shredding",
+        "microsoft-docs",
+        "microsoft-code-reference"
+      ],
       "never": ["azure-bicep-patterns", "azure-troubleshooting", "iac-common"]
     },
     "07b-Bicep Deploy": {
@@ -106,7 +122,7 @@
         "azure-diagrams",
         "session-resume"
       ],
-      "secondary": ["context-shredding"],
+      "secondary": ["context-shredding", "microsoft-docs"],
       "never": [
         "azure-bicep-patterns",
         "terraform-patterns",
@@ -116,7 +132,7 @@
     },
     "09-Diagnose": {
       "primary": ["azure-troubleshooting", "azure-defaults"],
-      "secondary": [],
+      "secondary": ["microsoft-docs", "microsoft-code-reference"],
       "never": [
         "azure-bicep-patterns",
         "terraform-patterns",
