+++ .github/workflows/azure-deprecation-tracker.yml
@@ -77,11 +77,11 @@ jobs:
             dependencies
         id: cpr
 
-      - name: Enable auto-merge on deprecation PR
+      - name: Log pull request
         if: steps.cpr.outputs.pull-request-number != ''
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: gh pr merge --auto --squash "${{ steps.cpr.outputs.pull-request-number }}"
+        run: |
+          echo "✅ Created/updated PR #${{ steps.cpr.outputs.pull-request-number }}"
+          echo "🔗 ${{ steps.cpr.outputs.pull-request-url }}"
 
       - name: Log deprecation summary
         if: steps.changes.outputs.changed == 'true'
