+++ .vscode/mcp.json
@@ -7,21 +7,18 @@
     "azure-pricing": {
       "type": "stdio",
       "command": "${workspaceFolder}/mcp/azure-pricing-mcp/.venv/bin/python",
-      "args": [
-        "-m",
-        "azure_pricing_mcp"
-      ],
+      "args": ["-m", "azure_pricing_mcp"],
       "cwd": "${workspaceFolder}/mcp/azure-pricing-mcp/src"
     },
     "terraform": {
       "type": "stdio",
       "command": "/go/bin/terraform-mcp-server",
-      "args": [
-        "stdio",
-        "--toolsets",
-        "registry"
-      ],
+      "args": ["stdio", "--toolsets", "registry"],
       "env": {}
+    },
+    "microsoft-learn": {
+      "type": "http",
+      "url": "https://learn.microsoft.com/api/mcp?maxTokenBudget=4000"
     }
   }
 }
