+++ docs/faq.md
@@ -85,7 +85,7 @@ Frequently asked questions about Agentic InfraOps.
     No. Agentic InfraOps requires:
 
     - **GitHub Copilot** — cloud-hosted AI service
-    - **MCP servers** — Azure Pricing, GitHub, and Terraform MCP servers
+    - **MCP servers** — Azure Pricing, GitHub, Terraform, and Microsoft Learn MCP servers
       provide real-time data to agents
 
     The dev container itself runs locally, but agent conversations and MCP tool calls
