+++ package.json
@@ -9,7 +9,6 @@
     "lint:md:fix": "markdownlint-cli2 --fix '**/*.md' '#node_modules'",
     "lint:links": "find . -name '*.md' ! -path './node_modules/*' ! -path './infra/*' | xargs -I {} markdown-link-check {} --config .markdown-link-check.json",
     "lint:links:docs": "find docs -name '*.md' | xargs -I {} markdown-link-check {} --config .markdown-link-check.json",
-    "lint:cost-estimate-templates": "echo '⚠️  Merged into lint:artifact-templates' && node scripts/validate-artifact-templates.mjs",
     "lint:artifact-templates": "node scripts/validate-artifact-templates.mjs",
     "lint:h2-sync": "node scripts/validate-h2-sync.mjs",
     "fix:artifact-h2": "node scripts/fix-artifact-h2.mjs",
@@ -31,13 +30,9 @@
     "lint:terraform-fmt": "terraform fmt -check -recursive infra/terraform/",
     "validate:terraform": "for dir in infra/terraform/*/; do [ -f \"$dir/main.tf\" ] && (cd \"$dir\" && terraform init -backend=false -input=false > /dev/null 2>&1 && terraform validate) || true; done",
     "validate:vscode": "node scripts/validate-vscode-config.mjs",
-    "validate:agents": "npm run lint:agent-frontmatter && npm run lint:skills-format",
-    "test:1109": "npm run validate:vscode && npm run validate:agents && echo '✅ VS Code 1.109 validation passed'",
     "fetch:deprecations": "node scripts/fetch-azure-deprecations.mjs",
     "snapshot:baseline": "bash scripts/snapshot-agent-context.sh",
     "diff:baseline": "bash scripts/diff-context-baseline.sh",
-    "lint:skill-size": "node scripts/validate-skill-size.mjs",
-    "lint:agent-body-size": "node scripts/validate-agent-body-size.mjs",
     "lint:glob-audit": "node scripts/validate-glob-audit.mjs",
     "lint:skill-references": "node scripts/validate-skill-references.mjs",
     "lint:orphaned-content": "node scripts/validate-orphaned-content.mjs",
@@ -48,7 +43,7 @@
     "validate:skill-digests": "node scripts/validate-skill-digests.mjs",
     "validate:docs-sync": "node scripts/validate-docs-sync.mjs",
     "init": "node scripts/init-from-template.mjs",
-    "validate:_node": "run-p lint:artifact-templates lint:h2-sync lint:version-sync lint:deprecated-refs lint:mcp-config lint:agent-frontmatter lint:skills-format lint:instruction-frontmatter lint:governance-refs validate:instruction-refs validate:session-state validate:session-lock validate:workflow-graph validate:agent-registry validate:skill-affinity validate:skill-digests lint:json lint:skill-size lint:agent-body-size lint:glob-audit lint:skill-references lint:orphaned-content lint:docs-freshness",
+    "validate:_node": "run-p lint:artifact-templates lint:h2-sync lint:version-sync lint:deprecated-refs lint:mcp-config lint:agent-frontmatter lint:skills-format lint:instruction-frontmatter lint:governance-refs validate:instruction-refs validate:session-state validate:session-lock validate:workflow-graph validate:agent-registry validate:skill-affinity validate:skill-digests lint:json lint:glob-audit lint:skill-references lint:orphaned-content lint:docs-freshness validate:docs-sync validate:vscode",
     "validate:_external": "run-p lint:md lint:python lint:terraform-fmt validate:terraform",
     "validate:all": "run-p --continue-on-error validate:_node validate:_external && echo '✅ All validations passed'"
   },
