+++ scripts/validate-instruction-frontmatter.mjs
@@ -10,38 +10,40 @@
  * node scripts/validate-instruction-frontmatter.mjs
  */
 
-import fs from "node:fs";
 import path from "node:path";
 import { getInstructions } from "./_lib/workspace-index.mjs";
+import { Reporter } from "./_lib/reporter.mjs";
 
-// Shared parser lowercases keys; match against lowercase allowed fields
 const REQUIRED_FIELDS = ["description", "applyto"];
 const REQUIRED_FIELDS_DISPLAY = ["description", "applyTo"];
 const OPTIONAL_FIELDS = ["name"];
 const ALLOWED_FIELDS = [...REQUIRED_FIELDS, ...OPTIONAL_FIELDS];
 const ALLOWED_FIELDS_DISPLAY = [...REQUIRED_FIELDS_DISPLAY, ...OPTIONAL_FIELDS];
 
-let errors = 0;
+const r = new Reporter("Instruction Frontmatter Validator");
+r.header();
 
-function validateFile(fileName, instr) {
-  const { path: filePath, content, frontmatter: fm } = instr;
+const instructions = getInstructions();
+
+console.log(`Found ${instructions.size} instruction file(s)\n`);
+
+for (const [fileName, instr] of instructions) {
+  r.tick();
+  const { path: filePath, frontmatter: fm } = instr;
   const relPath = path.relative(process.cwd(), filePath);
 
   if (!fm) {
-    console.log(
-      `::error file=${relPath},line=1::Missing YAML frontmatter (requires description and applyTo)`,
+    r.error(
+      relPath,
+      "Missing YAML frontmatter (requires description and applyTo)",
     );
-    errors++;
-    return;
+    continue;
   }
 
   for (const field of REQUIRED_FIELDS) {
     if (!fm[field]) {
       const display = REQUIRED_FIELDS_DISPLAY[REQUIRED_FIELDS.indexOf(field)];
-      console.log(
-        `::error file=${relPath},line=1::Missing required frontmatter field: ${display}`,
-      );
-      errors++;
+      r.error(relPath, `Missing required field: ${display}`);
     }
   }
 
@@ -49,27 +51,15 @@ function validateFile(fileName, instr) {
     (k) => !ALLOWED_FIELDS.includes(k),
   );
   if (unknownFields.length > 0) {
-    console.log(
-      `::error file=${relPath},line=1::Unknown frontmatter fields: ${unknownFields.join(", ")} (allowed: ${ALLOWED_FIELDS_DISPLAY.join(", ")})`,
+    r.error(
+      relPath,
+      `Unknown frontmatter fields: ${unknownFields.join(", ")} (allowed: ${ALLOWED_FIELDS_DISPLAY.join(", ")})`,
     );
-    errors++;
   }
 }
 
-console.log("🔍 Instruction File Frontmatter Validator\n");
-
-const instructions = getInstructions();
-
-console.log(`Found ${instructions.size} instruction file(s)\n`);
-
-for (const [fileName, instr] of instructions) {
-  validateFile(fileName, instr);
-}
-
-console.log(`\n${"=".repeat(50)}`);
-if (errors > 0) {
-  console.log(`❌ ${errors} error(s)`);
-  process.exit(1);
-} else {
-  console.log(`✅ All instruction files valid`);
-}
+r.summary();
+r.exitOnError(
+  "All instruction files valid",
+  "Instruction frontmatter validation FAILED",
+);
